+++ src/device/graphics/VESA.h
@@ -15,12 +15,12 @@
 #include "kernel/log/Logger.h"
 
 // Ausgewaehlte Grafikmodi mit Mode-Nummer
-constexpr const unsigned int MODE_640_480_16BITS = 0x111;
-constexpr const unsigned int MODE_640_480_24BITS = 0x112;
-constexpr const unsigned int MODE_800_600_16BITS = 0x114;
-constexpr const unsigned int MODE_800_600_24BITS = 0x115;
-constexpr const unsigned int MODE_1024_768_16BITS = 0x117;
-constexpr const unsigned int MODE_1024_768_24BITS = 0x118;
+constexpr const uint16_t MODE_640_480_16BITS = 0x111;
+constexpr const uint16_t MODE_640_480_24BITS = 0x112;
+constexpr const uint16_t MODE_800_600_16BITS = 0x114;
+constexpr const uint16_t MODE_800_600_24BITS = 0x115;
+constexpr const uint16_t MODE_1024_768_16BITS = 0x117;
+constexpr const uint16_t MODE_1024_768_24BITS = 0x118;
 
 class VESA : public LFBgraphics {
 private:
